extract component routesLimitItem
