+++ src/components/pages/routes-block/components/routes-header/routesHeader.module.sass
@@ -12,18 +12,7 @@
   width: fit-content
   outline: none
 
-.showedNums
+.limits
   display: inline-flex
   direction: row
-  gap: 0.625rem
-
-.showedNum
-  cursor: pointer
-  input
-    display: none
-
-    &:checked
-      + span
-        background-color: #fff
-        color: #000
-        font-weight: 700
+  gap: 0.5rem
