+++ src/components/pages/routes-block/components/routes-limit-item/routesLimitItem.module.sass
@@ -1,0 +1,10 @@
+.limit
+  cursor: pointer
+  input
+    display: none
+
+    &:checked
+      + span
+        background-color: #fff
+        color: #000
+        font-weight: 700
